docs/requirements.txt
aiohttp==3.8.6 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
alabaster==0.7.13 ; python_version >= "3.11" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.11" and python_version < "4.0"
astroid==2.15.8 ; python_version >= "3.11" and python_version < "4.0"
asttokens==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
async-lru==2.0.4 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
babel==2.13.0 ; python_version >= "3.11" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
contextvars==2.4 ; python_version >= "3.11" and python_version < "4.0"
contourpy==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.4.2 ; python_version >= "3.11" and python_version < "4.0"
docutils==0.20.1 ; python_version >= "3.11" and python_version < "4.0"
executing==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.104.0 ; python_version >= "3.11" and python_version < "4.0"
fonttools==4.43.1 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.4.0 ; python_version >= "3.11" and python_version < "4.0"
furo==2023.9.10 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.25.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
imagesize==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
immutables==0.20 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
jsonref==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.11" and python_version < "4.0"
lazy-object-proxy==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
matplotlib==3.8.0 ; python_version >= "3.11" and python_version < "4.0"
mdit-py-plugins==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
motor==3.3.1 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.11" and python_version < "4.0"
myst-parser==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.25.2 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.1.0 ; python_version >= "3.11" and python_version < "4.0"
pure-eval==0.2.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==1.10.13 ; python_version >= "3.11" and python_version < "4.0"
pydantic[dotenv]==1.10.13 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.16.1 ; python_version >= "3.11" and python_version < "4.0"
pymongo==4.5.0 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytube==15.0.0 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
sentry-sdk[fastapi,httpx,pure-eval,pymongo]==1.32.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools-scm==8.0.4 ; python_version >= "3.11" and python_version < "4.0"
setuptools==68.2.2 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
snowballstemmer==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.11" and python_version < "4.0"
sphinx-autodoc2 @ git+https://github.com/jag-k/sphinx-autodoc2.git@HEAD ; python_version >= "3.11" and python_version < "4.0"
sphinx-basic-ng==1.0.0b2 ; python_version >= "3.11" and python_version < "4.0"
sphinx-copybutton==0.5.2 ; python_version >= "3.11" and python_version < "4.0"
sphinx-inline-tabs==2023.4.21 ; python_version >= "3.11" and python_version < "4.0"
sphinx==7.1.1 ; python_version >= "3.11" and python_version < "4.0"
sphinxcontrib-applehelp==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
sphinxcontrib-devhelp==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
sphinxcontrib-htmlhelp==2.0.4 ; python_version >= "3.11" and python_version < "4.0"
sphinxcontrib-jsmath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
sphinxcontrib-qthelp==1.0.6 ; python_version >= "3.11" and python_version < "4.0"
sphinxcontrib-serializinghtml==1.1.9 ; python_version >= "3.11" and python_version < "4.0"
sphinxext-opengraph==0.8.2 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.0.7 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.23.2 ; python_version >= "3.11" and python_version < "4.0"
wrapt==1.15.0 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.11" and python_version < "4.0"
